// src/model/InformacaoMantis.ts
export default interface InformacaoMantis {
  categoria: string;
  codigoCliente: string;
  complexidade: string;
  crReact: string;
  dataEnvio: string;
  descricao: string;
  estado: string;
  informacaoAdicional: string;
  numeroCaso: string;
  prioridade: string;
  projeto: string;
  resumo: string;
  ultimaAtualizacao: string;
  versao: string;
}